:bug Lines now actually render when paused.
Notepad++ diffcheck mixed up the diff.

+++ Assets/Code/Camera/PlayerDebug.cs
@@ -2,12 +2,14 @@
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor;
+using System;
 
 public class PlayerDebug : MonoBehaviour {
 
 	public static PlayerDebug instance;
 	public Material mat;
-	public float circleSegments = 20;
+	public int circleSegments = 16;
+	public bool renderOnPause = true;
 	List<Vector3> verts;
 	List<Color> colors;
 	List<float> times;
@@ -16,7 +18,11 @@ public class PlayerDebug : MonoBehaviour {
 	List<float> circleRadii;
 	List<Color> circleColors;
 
+	bool paused = false;
+
 	void Awake(){
+		EditorApplication.playmodeStateChanged += OnStateChange;
+
 		instance = this;
 		verts = new List<Vector3>();
 		colors = new List<Color>();
@@ -26,6 +32,11 @@ public class PlayerDebug : MonoBehaviour {
 		circleRadii = new List<float>();
 		circleColors = new List<Color>();
 	}
+
+	private void OnStateChange(){
+		paused = renderOnPause && Application.isEditor && EditorApplication.isPaused;
+	}
+
 	// Use this for initialization
 	void Start () {
 
@@ -33,7 +44,7 @@ public class PlayerDebug : MonoBehaviour {
 
 	// Update is called once per frame
 	void Update () {
-
+		
 	}
 
 	public static void DrawCircle(Vector3 position, float radius, Color? color = null){
@@ -107,24 +118,26 @@ public class PlayerDebug : MonoBehaviour {
 		GL.End();
 		GL.PopMatrix();
 
+		if(!paused){
 
-
-		int count = times.Count;
-		//Add line back in if time isn't up
-		float t = Time.deltaTime;
-		for (int i = 0, timesCount = times.Count; i < timesCount; i++) {
-			float d = times [i];
-			if(d >= 0){
-				times.Add(d - t);
-				verts.Add(verts[i*2]);
-				verts.Add(verts[i*2 + 1]);
-				colors.Add(colors[i]);
+			int count = times.Count;
+			//Add line back in if time isn't up
+			float t = Time.deltaTime;
+			for (int i = 0, timesCount = times.Count; i < timesCount; i++) {
+				float d = times [i];
+				if(d >= 0){
+					times.Add(d - t);
+					verts.Add(verts[i*2]);
+					verts.Add(verts[i*2 + 1]);
+					colors.Add(colors[i]);
+				}
 			}
-		}
+   
 
-		times.RemoveRange(0,count);
-		verts.RemoveRange(0,count * 2);
-		colors.RemoveRange(0,count);
+			times.RemoveRange(0,count);
+			verts.RemoveRange(0,count * 2);
+			colors.RemoveRange(0,count);
+		}
 	}
 
 	void DrawCircles(){
@@ -135,6 +148,7 @@ public class PlayerDebug : MonoBehaviour {
 		mat.color = Color.white;
 
 		const float totalRadians = Mathf.PI * 2f;
+		circleSegments = Math.Max(3, circleSegments);
 		float angleInc = totalRadians / circleSegments;
 
 		for (int i = 0, circleCentersCount = circleCenters.Count; i < circleCentersCount; i++) {
@@ -163,6 +177,17 @@ public class PlayerDebug : MonoBehaviour {
 		}
 		GL.PopMatrix();
 
+						
+					  
+					   
 
+		if(!paused){
+			circleCenters.Clear();
+			circleRadii.Clear();
+			circleColors.Clear();
+		}
 	}
+
+
+
 }
